Fix migration: check if column exists before removing (#1141)

+++ db/migrate/20251211100000_add_web_search_content_and_timestamps_to_ai_report_sections.rb
@@ -11,7 +11,9 @@ class AddWebSearchContentAndTimestampsToAiReportSections < ActiveRecord::Migrati
     add_column :ai_report_sections, :created_at_web_included, :datetime
     add_column :ai_report_sections, :updated_at_web_included, :datetime
 
-    # Remove show_web_search as it will be derived from timestamps
-    remove_column :ai_report_sections, :show_web_search, :boolean, default: false
+    # Remove show_web_search as it will be derived from timestamps (only if it exists)
+    if column_exists?(:ai_report_sections, :show_web_search)
+      remove_column :ai_report_sections, :show_web_search, :boolean, default: false
+    end
   end
 end
